Refactor render function
