+++ src/index.html
@@ -3,7 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Wbpack Exercise</title>
+    <link rel="stylesheet" href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css" integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p" crossorigin="anonymous"/>
+    <title>JS Capstone Project</title>
 </head>
 <body>
     <header>
@@ -17,7 +18,10 @@
         </a>
     </nav>
     </header>
+  
     <main>
+        <section id="popup">
+        </section>
         <ul class="meals">
         </ul>
     </main>
